chapter_001/src/main/java/ru/job4j/it/EvenIterator.java
package ru.job4j.it;

import java.util.Iterator;
import java.util.NoSuchElementException;

public class EvenIterator implements Iterator<Integer> {
    private final int[] data;
    private int evenPoint = 0;

    public EvenIterator(int[] data) {
        this.data = data;
    }

    @Override
    public boolean hasNext() {
        boolean result = false;
        for (int i = evenPoint; i < data.length; i++) {
            if (data[i] % 2 == 0) {
                result = true;
                evenPoint = i;
                break;
            }
        }
        return result;
    }

    @Override
    public Integer next() {
        if (!hasNext()) {
            throw new NoSuchElementException();
        }
        return data[evenPoint++];
    }
}
